Refactor: Extract modules for improved architecture
Extract code from lib.rs into dedicated modules to improve maintainability:

- models/: BibItem and Citation data structures
- parser/: Bibliography parsing logic (nom-bibtex)
- renderer/: HTML generation from BibItems
- citation/: Citation replacement and per-chapter bibliography
- io/: File loading and Zotero download operations

Benefits:
- Reduced lib.rs from 770 lines to 202 lines (73% reduction)
- Clear separation of concerns with single responsibility per module
- Improved testability through proper module boundaries
- Zero functionality changes - all tests pass, builds successful
- Prepared architecture for backend abstraction

All 17 tests pass. No breaking changes to user-facing functionality.
