+++ src/lib.rs
@@ -1,26 +1,25 @@
-#[macro_use]
-extern crate lazy_static;
-
-use std::cell::RefCell;
-use std::collections::{HashMap, HashSet};
 use std::fs;
-use std::io::Read;
-use std::path::{Component, Path, PathBuf};
+use std::path::{Path, PathBuf};
 
-use crate::config::{Config, SortOrder};
 use anyhow::{anyhow, Context};
 use handlebars::Handlebars;
-use indexmap::IndexMap;
-use mdbook_preprocessor::book::{Book, BookItem, Chapter};
-use mdbook_preprocessor::errors::{Error, Result as MdResult};
+use mdbook_preprocessor::book::{Book, Chapter};
+use mdbook_preprocessor::errors::Error;
 use mdbook_preprocessor::{Preprocessor, PreprocessorContext};
-use nom_bibtex::*;
-use regex::Regex;
-use reqwest::blocking::Response;
-use serde::{Deserialize, Serialize};
 
+mod citation;
 mod config;
 mod file_utils;
+mod io;
+mod models;
+mod parser;
+mod renderer;
+
+use crate::config::Config;
+
+// Re-export for tests
+#[cfg(test)]
+pub use citation::{AT_REF_PATTERN, REF_PATTERN};
 
 static NAME: &str = "bib";
 static BIB_OUT_FILE: &str = "bibliography";
@@ -34,8 +33,6 @@ impl Default for Bibliography {
 }
 
 impl Bibliography {
-    // Get references and info from the bibliography file specified in the config
-    // arg "bibliography" or in Zotero
     fn retrieve_bibliography_content(
         ctx: &PreprocessorContext,
         cfg: &Config,
@@ -46,12 +43,10 @@ impl Bibliography {
                 let mut biblio_path = ctx.root.join(&ctx.config.book.src);
                 biblio_path = biblio_path.join(Path::new(&biblio_file));
                 if !biblio_path.exists() {
-                    Err(anyhow!(format!(
-                        "Bibliography file {biblio_path:?} not found!",
-                    )))
+                    Err(anyhow!("Bibliography file {biblio_path:?} not found!",))
                 } else {
                     tracing::info!("Bibliography path: {}", biblio_path.display());
-                    load_bibliography(biblio_path)
+                    io::load_bibliography(biblio_path)
                 }
             }
             _ => {
@@ -59,14 +54,13 @@ impl Bibliography {
                 match &cfg.zotero_uid {
                     Some(uid) => {
                         let user_id = uid.to_string();
-                        let bib_str = download_bib_from_zotero(user_id).unwrap_or_default();
+                        let bib_str = io::download_bib_from_zotero(user_id).unwrap_or_default();
                         if !bib_str.is_empty() {
                             let biblio_path = ctx.root.join(Path::new("my_zotero.bib"));
                             tracing::info!("Saving Zotero bibliography to {:?}", biblio_path);
                             let _ = fs::write(biblio_path, &bib_str);
                             Ok(bib_str)
                         } else {
-                            // warn!("Bib content retrieved from Zotero is empty!");
                             Err(anyhow!("Bib content retrieved from Zotero is empty!"))
                         }
                     }
@@ -75,83 +69,6 @@ impl Bibliography {
             }
         };
         bib_content
-    }
-
-    fn generate_bibliography_html(
-        bibliography: &IndexMap<String, BibItem>,
-        cited: &HashSet<String>,
-        cited_only: bool,
-        handlebars: &Handlebars,
-        order: SortOrder,
-    ) -> String {
-        let sorted: Vec<(&str, &BibItem)> = match order {
-            SortOrder::None => bibliography.iter().map(|(k, v)| (k.as_str(), v)).collect(),
-            SortOrder::Key => {
-                let mut v: Vec<(&str, &BibItem)> =
-                    bibliography.iter().map(|(k, v)| (k.as_str(), v)).collect();
-                v.sort_by_key(|item| item.0);
-                v
-            }
-            SortOrder::Author => {
-                let empty = "!".to_string();
-                let mut v: Vec<(&str, &BibItem)> =
-                    bibliography.iter().map(|(k, v)| (k.as_str(), v)).collect();
-                v.sort_by_cached_key(|item| {
-                    let val: &str = item
-                        .1
-                        .authors
-                        .first()
-                        .map(|vec| vec.first().unwrap_or(&empty))
-                        .unwrap_or(&empty);
-                    val
-                });
-                v
-            }
-            SortOrder::Index => {
-                let mut v: Vec<(&str, &BibItem)> =
-                    bibliography.iter().map(|(k, v)| (k.as_str(), v)).collect();
-                v.sort_by_key(|item| item.1.index);
-                v
-            }
-        };
-
-        let mut content: String = String::from("");
-        for (key, value) in sorted {
-            if !cited_only || cited.contains(key) {
-                content.push_str(handlebars.render("references", &value).unwrap().as_str());
-            }
-        }
-
-        tracing::debug!("Generated Bib Content: {:?}", content);
-        content
-    }
-
-    fn expand_cite_references_in_book(
-        book: &mut Book,
-        bibliography: &mut IndexMap<String, BibItem>,
-        handlebars: &Handlebars,
-    ) -> HashSet<String> {
-        let mut cited = HashSet::new();
-        let mut last_index = 0;
-        book.for_each_mut(|section: &mut BookItem| {
-            if let BookItem::Chapter(ref mut ch) = *section {
-                if let Some(ref chapter_path) = ch.path {
-                    tracing::info!(
-                        "Replacing placeholders: {{#cite ...}} and @@citation in {}",
-                        chapter_path.as_path().display()
-                    );
-                    let new_content = replace_all_placeholders(
-                        ch,
-                        bibliography,
-                        &mut cited,
-                        handlebars,
-                        &mut last_index,
-                    );
-                    ch.content = new_content;
-                }
-            }
-        });
-        cited
     }
 
     fn create_bibliography_chapter(
@@ -173,308 +90,6 @@ impl Bibliography {
             PathBuf::from(format!("{BIB_OUT_FILE}.md")),
             Vec::new(),
         )
-    }
-}
-
-/// Bibliography item representation.
-/// TODO: Complete with more fields when necessary
-#[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
-pub struct BibItem {
-    /// The citation key.
-    pub citation_key: String,
-    /// The article's title.
-    pub title: String,
-    /// The article's author/s.
-    pub authors: Vec<Vec<String>>,
-    /// Pub month.
-    pub pub_month: String,
-    /// Pub year.
-    pub pub_year: String,
-    /// Summary/Abstract.
-    pub summary: String,
-    /// The article's url.
-    pub url: Option<String>,
-    /// The item's index for first citation in the book.
-    pub index: Option<u32>,
-}
-
-impl BibItem {
-    /// Create a new bib item with the provided content.
-    pub fn new(
-        citation_key: &str,
-        title: String,
-        authors: Vec<Vec<String>>,
-        pub_month: String,
-        pub_year: String,
-        summary: String,
-        url: Option<String>,
-    ) -> BibItem {
-        BibItem {
-            citation_key: citation_key.to_string(),
-            title,
-            authors,
-            pub_month,
-            pub_year,
-            summary,
-            url,
-            index: None,
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
-struct Citation {
-    pub item: BibItem,
-    pub path: String,
-}
-
-/// Load bibliography from file
-pub(crate) fn load_bibliography<P: AsRef<Path>>(biblio_file: P) -> MdResult<String> {
-    tracing::info!("Loading bibliography from {:?}...", biblio_file.as_ref());
-
-    let biblio_file_ext = file_utils::get_filename_extension(biblio_file.as_ref());
-    if biblio_file_ext.unwrap_or_default().to_lowercase() != "bib" {
-        tracing::warn!(
-            "Only biblatex-based bibliography is supported for now! Yours: {:?}",
-            biblio_file.as_ref()
-        );
-        return Ok("".to_string());
-    }
-    Ok(fs::read_to_string(biblio_file)?)
-}
-
-fn extract_biblio_data_and_link_info(res: &mut Response) -> (String, String) {
-    let mut biblio_chunk = String::new();
-    let _ = res.read_to_string(&mut biblio_chunk);
-    let link_info_in_header = res.headers().get("link");
-    tracing::debug!("Header Link content: {:?}", link_info_in_header);
-    let link_info_as_str = link_info_in_header.unwrap().to_str();
-
-    (link_info_as_str.unwrap().to_string(), biblio_chunk)
-}
-
-/// Download bibliography from Zotero
-pub(crate) fn download_bib_from_zotero(user_id: String) -> MdResult<String, Error> {
-    let mut url = format!("https://api.zotero.org/users/{user_id}/items?format=biblatex&style=biblatex&limit=100&sort=creator&v=3");
-    tracing::info!("Zotero's URL biblio source:\n{url:?}");
-    let mut res = reqwest::blocking::get(&url)?;
-    if res.status().is_client_error() || res.status().is_client_error() {
-        Err(anyhow!(format!(
-            "Error accessing Zotero API {:?}",
-            res.error_for_status()
-        )))
-    } else {
-        let (mut link_str, mut bib_content) = extract_biblio_data_and_link_info(&mut res);
-        while link_str.contains("next") {
-            // Extract next chunk URL
-            let next_idx = link_str.find("rel=\"next\"").unwrap();
-            let end_bytes = next_idx - 3; // The > of the "next" link is 3 chars before rel=\"next\" pattern
-            let slice = &link_str[..end_bytes];
-            let start_bytes = slice.rfind('<').unwrap_or(0);
-            url = link_str[(start_bytes + 1)..end_bytes].to_string();
-            tracing::info!("Next biblio chunk URL:\n{:?}", url);
-            res = reqwest::blocking::get(&url)?;
-            let (new_link_str, new_bib_part) = extract_biblio_data_and_link_info(&mut res);
-            link_str = new_link_str;
-            bib_content.push_str(&new_bib_part);
-        }
-        Ok(bib_content)
-    }
-}
-
-/// Gets the references and info created from bibliography.
-pub(crate) fn build_bibliography(
-    raw_content: String,
-) -> MdResult<IndexMap<String, BibItem>, Error> {
-    tracing::info!("Building bibliography...");
-
-    // Filter quotes (") that may appear in abstracts, etc. and that Bibtex parser doesn't like
-    let mut biblatex_content = raw_content.replace('\"', "");
-    // Expressions in the content such as R@10 are not parsed well
-    let re = Regex::new(r" (?P<before>[A-Za-z])@(?P<after>\d+) ").unwrap();
-    biblatex_content = re
-        .replace_all(&biblatex_content, " ${before}_at_${after} ")
-        .into_owned();
-
-    tracing::info!("Attempting to parse BibTeX content...");
-    let bib = match Bibtex::parse(&biblatex_content) {
-        Ok(bib) => {
-            tracing::info!("Successfully parsed BibTeX content");
-            bib
-        }
-        Err(e) => {
-            tracing::error!("Failed to parse BibTeX content: {}", e);
-            tracing::error!("This might be due to malformed BibTeX syntax, missing braces, or invalid characters");
-            return Err(anyhow!("BibTeX parsing failed: {e}"));
-        }
-    };
-
-    let biblio = bib.bibliographies();
-    tracing::info!("{} bibliography items read", biblio.len());
-
-    let bibliography: IndexMap<String, BibItem> = biblio
-        .iter()
-        .map(|bib| {
-            let citation_key = bib.citation_key().to_string();
-            tracing::info!("Processing bibliography entry: {}", citation_key);
-
-            let tm: HashMap<String, String> = bib
-                .tags()
-                .iter()
-                .map(|(k, v)| (k.clone(), v.clone()))
-                .collect();
-
-            // Process authors with explicit error handling
-            let authors_str = match tm.get("author") {
-                Some(author) => {
-                    let mut clean_author = author.to_string();
-                    clean_author.retain(|c| c != '\n');
-                    tracing::debug!("Entry {}: author field = '{}'", citation_key, clean_author);
-                    clean_author
-                }
-                None => {
-                    tracing::warn!("Entry {}: missing author field, using 'N/A'", citation_key);
-                    "N/A".to_string()
-                }
-            };
-
-            // Process title with explicit error handling
-            let title = match tm.get("title") {
-                Some(title_val) => {
-                    tracing::debug!("Entry {}: title field = '{}'", citation_key, title_val);
-                    title_val.to_string()
-                }
-                None => {
-                    tracing::warn!(
-                        "Entry {}: missing title field, using 'Not Found'",
-                        citation_key
-                    );
-                    "Not Found".to_string()
-                }
-            };
-
-            // Process abstract/summary with explicit error handling
-            let summary = match tm.get("abstract") {
-                Some(abstract_val) => {
-                    tracing::debug!(
-                        "Entry {}: abstract field = '{}'",
-                        citation_key,
-                        abstract_val
-                    );
-                    abstract_val.to_string()
-                }
-                None => {
-                    tracing::debug!(
-                        "Entry {}: missing abstract field, using 'N/A'",
-                        citation_key
-                    );
-                    "N/A".to_string()
-                }
-            };
-
-            // Process URL with explicit error handling
-            let url: Option<String> = match tm.get("url") {
-                Some(url_val) => match url_val.parse::<String>() {
-                    Ok(parsed_url) => {
-                        tracing::debug!("Entry {}: url field = '{}'", citation_key, parsed_url);
-                        Some(parsed_url)
-                    }
-                    Err(e) => {
-                        tracing::warn!(
-                            "Entry {}: failed to parse URL '{}': {}",
-                            citation_key,
-                            url_val,
-                            e
-                        );
-                        None
-                    }
-                },
-                None => {
-                    tracing::debug!("Entry {}: missing url field", citation_key);
-                    None
-                }
-            };
-
-            // Process date with explicit error handling
-            let (pub_year, pub_month) = extract_date(&tm);
-            tracing::debug!(
-                "Entry {}: date fields = year='{}', month='{}'",
-                citation_key,
-                pub_year,
-                pub_month
-            );
-
-            // Process authors list with explicit error handling
-            let and_split = Regex::new(r"\band\b").expect("Broken regex");
-            let splits = and_split.split(&authors_str);
-            let authors: Vec<Vec<String>> = splits
-                .map(|a| {
-                    let author_parts: Vec<String> =
-                        a.trim().split(',').map(|b| b.trim().to_string()).collect();
-                    tracing::debug!("Entry {}: author part = '{:?}'", citation_key, author_parts);
-                    author_parts
-                })
-                .collect();
-
-            tracing::debug!(
-                "Entry {}: final authors list = '{:?}'",
-                citation_key,
-                authors
-            );
-
-            (
-                citation_key.clone(),
-                BibItem {
-                    citation_key,
-                    title,
-                    authors,
-                    pub_month,
-                    pub_year,
-                    summary,
-                    url,
-                    index: None,
-                },
-            )
-        })
-        .collect();
-    tracing::debug!("Bibiography content:\n{:?}", bibliography);
-
-    Ok(bibliography)
-}
-
-fn extract_date(tm: &HashMap<String, String>) -> (String, String) {
-    if let Some(date_str) = tm.get("date") {
-        tracing::debug!("Processing date field: '{}'", date_str);
-        let mut date = date_str.split('-');
-        let year = date.next().unwrap_or("N/A").to_string();
-        let month = date
-            .next()
-            .unwrap_or_else(|| tm.get("month").map(|s| s.as_str()).unwrap_or("N/A"))
-            .to_string();
-        tracing::debug!(
-            "Extracted from date field: year='{}', month='{}'",
-            year,
-            month
-        );
-        (year, month)
-    } else {
-        tracing::debug!("No date field found, looking for separate year/month fields");
-        let year = tm
-            .get("year")
-            .map(|s| s.as_str())
-            .unwrap_or("N/A")
-            .to_string();
-        let month = tm
-            .get("month")
-            .map(|s| s.as_str())
-            .unwrap_or("N/A")
-            .to_string();
-        tracing::debug!(
-            "Extracted from separate fields: year='{}', month='{}'",
-            year,
-            month
-        );
-        (year, month)
     }
 }
 
@@ -507,6 +122,7 @@ impl Preprocessor for Bibliography {
                 return Ok(book);
             }
         };
+
         // Configure template registry
         let mut handlebars = Handlebars::new();
         handlebars
@@ -535,7 +151,7 @@ impl Preprocessor for Bibliography {
             return Ok(book);
         }
 
-        let bibliography = build_bibliography(bib_content?);
+        let bibliography = parser::parse_bibliography(bib_content?);
         if bibliography.is_err() {
             tracing::warn!(
                 "Error building Bibliography from raw content. Skipping render: {:?}",
@@ -547,12 +163,17 @@ impl Preprocessor for Bibliography {
         let mut bib = bibliography.unwrap();
 
         if config.add_bib_in_each_chapter {
-            add_bib_at_end_of_chapters(&mut book, &mut bib, &handlebars, config.order.to_owned());
+            citation::add_bib_at_end_of_chapters(
+                &mut book,
+                &mut bib,
+                &handlebars,
+                config.order.clone(),
+            );
         }
 
-        let cited = Bibliography::expand_cite_references_in_book(&mut book, &mut bib, &handlebars);
+        let cited = citation::expand_cite_references_in_book(&mut book, &mut bib, &handlebars);
 
-        let bib_content_html = Bibliography::generate_bibliography_html(
+        let bib_content_html = renderer::generate_bibliography_html(
             &bib,
             &cited,
             config.cited_only,
@@ -575,184 +196,6 @@ impl Preprocessor for Bibliography {
     fn supports_renderer(&self, renderer: &str) -> Result<bool, anyhow::Error> {
         Ok(renderer != "not-supported")
     }
-}
-
-fn add_bib_at_end_of_chapters(
-    book: &mut Book,
-    bibliography: &mut IndexMap<String, BibItem>,
-    handlebars: &Handlebars,
-    order: SortOrder,
-) {
-    use regex::Regex;
-
-    use std::collections::HashSet;
-
-    lazy_static! {
-        static ref REF_REGEX: Regex = Regex::new(REF_PATTERN).unwrap();
-        static ref AT_REF_REGEX: Regex = Regex::new(AT_REF_PATTERN).unwrap();
-    }
-
-    book.for_each_mut(|section: &mut BookItem| {
-        if let BookItem::Chapter(ref mut ch) = *section {
-            if let Some(ref chapter_path) = ch.path {
-                tracing::info!(
-                    "Adding bibliography at the end of chapter {}",
-                    chapter_path.as_path().display()
-                );
-
-                let mut cited = HashSet::new();
-                // Find all {{#cite ...}} keys
-                for caps in REF_REGEX.captures_iter(&ch.content) {
-                    if let Some(cite) = caps.get(1) {
-                        cited.insert(cite.as_str().trim().to_owned());
-                    }
-                }
-                // Find all @@... keys
-                for caps in AT_REF_REGEX.captures_iter(&ch.content) {
-                    if let Some(cite) = caps.get(2) {
-                        cited.insert(cite.as_str().trim().to_owned());
-                    }
-                }
-                tracing::info!("Refs cited in this chapter: {:?}", cited);
-
-                let ch_bib_header_html = handlebars
-                    .render("chapter_refs", &String::new())
-                    .unwrap()
-                    .as_str()
-                    .to_string();
-
-                let ch_bib_content_html = Bibliography::generate_bibliography_html(
-                    bibliography,
-                    &cited,
-                    true,
-                    handlebars,
-                    order.clone(),
-                );
-
-                let new_content = String::from(ch.content.as_str())
-                    + ch_bib_header_html.as_str()
-                    + ch_bib_content_html.as_str();
-                ch.content = new_content;
-            }
-        }
-    });
-}
-
-fn replace_all_placeholders(
-    chapter: &Chapter,
-    bibliography: &mut IndexMap<String, BibItem>,
-    cited: &mut HashSet<String>,
-    handlebars: &Handlebars,
-    last_index: &mut u32,
-) -> String {
-    let chapter_path = chapter
-        .path
-        .as_deref()
-        .unwrap_or_else(|| std::path::Path::new(""));
-
-    lazy_static! {
-        static ref REF_REGEX: Regex = Regex::new(REF_PATTERN).unwrap();
-        static ref AT_REF_REGEX: Regex = Regex::new(AT_REF_PATTERN).unwrap();
-    }
-
-    // Wrap mutable state in RefCell for interior mutability
-    let bib = RefCell::new(bibliography);
-    let cited_set = RefCell::new(cited);
-    let idx = RefCell::new(last_index);
-
-    // First replace all {{#cite ...}}
-    let replaced = REF_REGEX.replace_all(&chapter.content, |caps: &regex::Captures| {
-        let cite = caps.get(1).map(|m| m.as_str().trim()).unwrap_or("");
-        cited_set.borrow_mut().insert(cite.to_owned());
-        let mut bib_mut = bib.borrow_mut();
-        let mut idx_mut = idx.borrow_mut();
-        if bib_mut.contains_key(cite) {
-            let path_to_root = breadcrumbs_up_to_root(chapter_path);
-            let item = bib_mut.get_mut(cite).unwrap();
-            if item.index.is_none() {
-                **idx_mut += 1;
-                item.index = Some(**idx_mut);
-            }
-            let citation = Citation {
-                item: item.to_owned(),
-                path: format!("{path_to_root}{BIB_OUT_FILE}.html"),
-            };
-            handlebars
-                .render("citation", &citation)
-                .unwrap()
-                .as_str()
-                .to_string()
-        } else {
-            format!("\\[Unknown bib ref: {cite}\\]")
-        }
-    });
-
-    // Then replace all @@cite
-    let replaced = AT_REF_REGEX.replace_all(&replaced, |caps: &regex::Captures| {
-        let cite = caps.get(2).map(|m| m.as_str().trim()).unwrap_or("");
-        cited_set.borrow_mut().insert(cite.to_owned());
-        let mut bib_mut = bib.borrow_mut();
-        let mut idx_mut = idx.borrow_mut();
-        if bib_mut.contains_key(cite) {
-            let path_to_root = breadcrumbs_up_to_root(chapter_path);
-            let item = bib_mut.get_mut(cite).unwrap();
-            if item.index.is_none() {
-                **idx_mut += 1;
-                item.index = Some(**idx_mut);
-            }
-            let citation = Citation {
-                item: item.to_owned(),
-                path: format!("{path_to_root}{BIB_OUT_FILE}.html"),
-            };
-            handlebars
-                .render("citation", &citation)
-                .unwrap()
-                .as_str()
-                .to_string()
-        } else {
-            format!("\\[Unknown bib ref: {cite}\\]")
-        }
-    });
-
-    replaced.into_owned()
-}
-
-// Regex patterns for citation placeholders
-// BibLaTeX-compliant character class for citation keys:
-// - Allowed: alphanumeric, underscore, hyphen, colon, dot, slash, at-symbol
-// - Forbidden: spaces, comma, quotes, hash, braces, percent, tilde, parentheses, equals
-pub(crate) const REF_PATTERN: &str = r"
-(?x)                       # insignificant whitespace mode
-\\\{\{\#.*\}\}               # match escaped placeholder
-|                            # or
-\{\{\s*                      # placeholder opening parens and whitespace
-\#cite                       # explicitly match #cite (only, not other mdBook helpers like #include, #title)
-\s+                          # separating whitespace
-([a-zA-Z0-9_\-:./@]+)        # citation key (capture group 1) - BibLaTeX compliant
-\s*\}\}                      # whitespace and placeholder closing parens";
-
-pub(crate) const AT_REF_PATTERN: &str = r##"(@@)([a-zA-Z0-9_\-/@]+(?:[.:][a-zA-Z0-9_\-/@]+)*)"##;
-
-fn breadcrumbs_up_to_root(source_file: &std::path::Path) -> String {
-    if source_file.as_os_str().is_empty() {
-        return "".into();
-    }
-
-    let components_count = source_file.components().fold(0, |acc, c| match c {
-        Component::Normal(_) => acc + 1,
-        Component::ParentDir => acc - 1,
-        Component::CurDir => acc,
-        Component::RootDir | Component::Prefix(_) => panic!(
-            "mdBook is not supposed to give us absolute paths, only relative from the book root."
-        ),
-    }) - 1;
-
-    let mut to_root = vec![".."; components_count].join("/");
-    if components_count > 0 {
-        to_root.push('/');
-    }
-
-    to_root
 }
 
 #[cfg(test)]
